feat: added errors

+++ controllers/movies.js
@@ -1,5 +1,9 @@
 const Movie = require('../models/movie');
 
+const NotFoundError = require('../errors/not-found-err');
+const ForbiddenError = require('../errors/forbidden-err');
+const BadRequestError = require('../errors/bad-request-err');
+
 module.exports.getMovies = async (req, res, next) => {
   try {
     const movies = await Movie.find({});
@@ -41,28 +45,28 @@ module.exports.createMovie = async (req, res, next) => {
     });
     res.status(201).send(movie);
   } catch (err) {
-    // if (err.name === 'ValidationError') {
-    //   next(new BadRequestError('Переданы некорректные данные при создании карточки.'));
-    //   return;
-    // }
+    if (err.name === 'ValidationError') {
+      next(new BadRequestError('Переданы некорректные данные при создании фильма.'));
+      return;
+    }
     next(err);
   }
 };
 
 module.exports.deleteMovie = async (req, res, next) => {
   try {
-    // const movie = await Movie.findById(req.params.cardId);
-    // if (!movie) throw new NotFoundError('Карточка с указанным _id не найдена.');
+    const movie = await Movie.findById(req.params.cardId);
+    if (!movie) throw new NotFoundError('Фильм с указанным id не найдена.');
 
-    // if (movie.owner.toString() !== req.user._id) throw new ForbiddenError('Чужая карточка.');
+    if (movie.owner.toString() !== req.user._id) throw new ForbiddenError('Чужой фильм.');
 
     const deletedMovie = await Movie.findByIdAndDelete(req.params.cardId);
     res.send(deletedMovie);
   } catch (err) {
-    // if (err.name === 'CastError') {
-    //   next(new BadRequestError('Передан некорректный _id карточки.'));
-    //   return;
-    // }
+    if (err.name === 'CastError') {
+      next(new BadRequestError('Передан некорректный id фильма.'));
+      return;
+    }
     next(err);
   }
 };
